Lift state up from MapContainer
Also move TrackingContainer from MapContainer into Router.
