+++ client/src/map/marker.tsx
@@ -63,8 +63,10 @@ export const SharedLocationMarker = styled(NonUserMarker)`
   background-color: green;
 `;
 
+export type PinKind = 'configure' | 'show' | 'none';
+
 interface PinProps {
-  type: 'configure' | 'show' | 'none';
+  type: PinKind;
   coords: Location;
   onClick(a: MouseEvent<HTMLButtonElement>): void;
 }
