+++ client/src/router.tsx
@@ -2,12 +2,14 @@ import React, { useState, useEffect } from 'react';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import styled from 'styled-components';
 import { Transition } from 'react-spring/renderprops';
+
+import TrackingContainer from './map/trackingContainer';
 import AboutContainer from './aboutContainer';
 import {
   GenuineBusContainer,
   MockBusContainer,
 } from './3dVisualisation/screenContainer';
-import UrlPromptContainer from './location/urlPromptContainer';
+import UrlPromptContainer, { MQTT_URL } from './location/urlPromptContainer';
 import { apiRoot } from './common/environment';
 import MapContainer from './map/mapContainer';
 import AdminPanel, { RaspberryLocation } from './admin/adminPanel';
@@ -18,6 +20,9 @@ import AuthApi, { Admin } from './admin/authApi';
 import ShareLocationModal from './map/shareLocationModal';
 import PublicShareModal from './map/publicShareModal';
 import { parseQuery, MapLocationQueryDecoder } from './common/urlParse';
+import { useUbiMqtt } from './location/mqttConnection';
+import { BeaconGeoLocation } from './location/mqttDeserialize';
+import { PinKind } from './map/marker';
 
 const NotFound = () => <h3>404 page not found</h3>;
 
@@ -49,10 +54,31 @@ const Router = () => {
   const [publicShareOpen, openPublicShare] = useState(false);
   const [bluetoothName, setBluetoothName] = useState<string | null>(null);
 
+  /**
+   * Used when user selects "only current" from the location prompt.
+   */
+  const [staticLocations, setStaticLocations] = useState<BeaconGeoLocation[]>(
+    []
+  );
+
   const queryParams = parseQuery(
     MapLocationQueryDecoder,
     document.location.search
   );
+
+  const fromQuery = !!(queryParams && queryParams.lat && queryParams.lon);
+  const initialPinType = fromQuery ? 'show' : 'none';
+
+  const [pinType, setPinType] = useState<PinKind>(initialPinType);
+
+  const mqttHost =
+    queryParams && queryParams.host ? queryParams.host : MQTT_URL;
+  const { beacons, lastKnownPosition } = useUbiMqtt(
+    mqttHost,
+    bluetoothName,
+    queryParams && queryParams.topic ? queryParams.topic : undefined
+  );
+
   const [nameModalOpen, setNameModalOpen] = useState(
     queryParams && queryParams.lat ? true : false
   );
@@ -86,6 +112,22 @@ const Router = () => {
           }}
           onClose={() => openPublicShare(false)}
           isOpen={publicShareOpen}
+        />
+      )}
+      {nameModalOpen && (
+        <TrackingContainer
+          beacons={beacons}
+          onClose={() => setNameModalOpen(false)}
+          confirmName={name => {
+            setBluetoothName(name);
+            setStaticLocations([]);
+            setPinType('none');
+            setNameModalOpen(false);
+          }}
+          onStaticSelected={name => {
+            const targetBeacons = beacons.filter(b => b.beaconId === name);
+            setStaticLocations(targetBeacons);
+          }}
         />
       )}
       <Fullscreen>
@@ -168,10 +210,13 @@ const Router = () => {
               render={props => (
                 <MapContainer
                   {...props}
-                  nameModalOpen={nameModalOpen}
+                  beacons={beacons}
+                  pinType={pinType}
+                  setPinType={setPinType}
+                  lastKnownPosition={lastKnownPosition}
+                  staticLocations={staticLocations}
                   setNameModalOpen={setNameModalOpen}
                   bluetoothName={bluetoothName}
-                  setBluetoothName={setBluetoothName}
                   roomReserved={roomReserved}
                   devices={devices}
                   getDeviceLocation={getDeviceLocation}
